refactor: Extract mesh_tools.py tabs into separate mixin modules
Splits the monolithic mesh_tools.py (2203 lines) into composable modules:
- mesh_tools.py: Main panel (428 lines) - shared infrastructure
- mesh_tools_meshbot.py: MeshBot tab (781 lines)
- mesh_tools_nodemap.py: Node Map tab (448 lines)
- mesh_tools_diagnostics.py: Diagnostics tab (429 lines)
- mesh_tools_sensors.py: Sensors tab (279 lines)

All files now under 1500 line threshold per v1.0 roadmap Phase 1.
Uses mixin pattern for clean composition while maintaining
full backwards compatibility. All 1297 tests pass.

Part of Phase 1: Stability for v1.0 roadmap.
