+++ src/gtk_ui/panels/mesh_tools.py
@@ -5,8 +5,16 @@ Combines:
 - MeshBot (BBS, games, inventory, weather)
 - Node Map (network visualization)
 - Diagnostics (health monitoring, event log)
+- Sensors (telemetry display)
 
 With shared resizable log output at bottom.
+
+Refactored to use tab mixins for maintainability (v0.4.8).
+Each tab is implemented in its own module:
+- mesh_tools_meshbot.py
+- mesh_tools_nodemap.py
+- mesh_tools_diagnostics.py
+- mesh_tools_sensors.py
 """
 
 import gi
@@ -16,7 +24,6 @@ from gi.repository import Gtk, Adw, GLib
 import threading
 import subprocess
 import os
-import json
 from pathlib import Path
 from datetime import datetime
 
@@ -41,18 +48,6 @@ except ImportError:
 # Import path utilities
 from utils.paths import get_real_user_home
 
-# Import meshtastic connection utilities
-try:
-    from utils.meshtastic_connection import (
-        MESHTASTIC_CONNECTION_LOCK,
-        wait_for_cooldown,
-        safe_close_interface
-    )
-    HAS_MESHTASTIC_LOCK = True
-except ImportError:
-    HAS_MESHTASTIC_LOCK = False
-    MESHTASTIC_CONNECTION_LOCK = None
-
 # Import settings manager
 try:
     from utils.common import SettingsManager
@@ -69,20 +64,27 @@ try:
 except ImportError:
     HAS_DIAGNOSTICS = False
 
-# Import service availability checker
-try:
-    from utils.service_check import check_port, check_service, require_service
-    HAS_SERVICE_CHECK = True
-except ImportError:
-    HAS_SERVICE_CHECK = False
+# Import tab mixins
+from .mesh_tools_meshbot import MeshBotTabMixin
+from .mesh_tools_nodemap import NodeMapTabMixin
+from .mesh_tools_diagnostics import DiagnosticsTabMixin
+from .mesh_tools_sensors import SensorsTabMixin
 
 
-class MeshToolsPanel(Gtk.Box):
+class MeshToolsPanel(
+    MeshBotTabMixin,
+    NodeMapTabMixin,
+    DiagnosticsTabMixin,
+    SensorsTabMixin,
+    Gtk.Box
+):
     """
     Consolidated mesh network tools panel.
 
-    Provides sub-tabs for MeshBot, Node Map, and Diagnostics
+    Provides sub-tabs for MeshBot, Node Map, Diagnostics, and Sensors
     with a shared resizable log output at the bottom.
+
+    Tab implementations are provided by mixin classes for maintainability.
     """
 
     SETTINGS_DEFAULTS = {
@@ -119,6 +121,10 @@ class MeshToolsPanel(Gtk.Box):
         # Initial status check
         self._schedule_timer(500, self._check_all_status)
 
+    # =========================================================================
+    # Timer Management
+    # =========================================================================
+
     def _schedule_timer(self, delay_ms: int, callback, *args) -> int:
         """Schedule a timer and track it for cleanup."""
         if args:
@@ -137,11 +143,19 @@ class MeshToolsPanel(Gtk.Box):
                 pass
         self._pending_timers.clear()
 
+    # =========================================================================
+    # Settings Management
+    # =========================================================================
+
     def _save_settings(self):
         """Save settings"""
         if HAS_SETTINGS:
             self._settings_mgr.update(self._settings)
             self._settings_mgr.save()
+
+    # =========================================================================
+    # UI Building
+    # =========================================================================
 
     def _build_ui(self):
         """Build the main UI with paned layout"""
@@ -171,7 +185,7 @@ class MeshToolsPanel(Gtk.Box):
         self._notebook = Gtk.Notebook()
         self._notebook.set_tab_pos(Gtk.PositionType.TOP)
 
-        # Add tabs
+        # Add tabs from mixins
         self._add_meshbot_tab()
         self._add_map_tab()
         self._add_diagnostics_tab()
@@ -260,740 +274,7 @@ class MeshToolsPanel(Gtk.Box):
         self._save_settings()
 
     # =========================================================================
-    # MeshBot Tab
-    # =========================================================================
-
-    def _add_meshbot_tab(self):
-        """Add MeshBot management tab"""
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=15)
-        box.set_margin_start(15)
-        box.set_margin_end(15)
-        box.set_margin_top(15)
-        box.set_margin_bottom(15)
-
-        # Status section
-        status_frame = Gtk.Frame()
-        status_frame.set_label("MeshBot Status")
-        status_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        status_box.set_margin_start(15)
-        status_box.set_margin_end(15)
-        status_box.set_margin_top(10)
-        status_box.set_margin_bottom(10)
-
-        # Status row
-        status_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=15)
-
-        self._meshbot_status_icon = Gtk.Image.new_from_icon_name("emblem-question-symbolic")
-        self._meshbot_status_icon.set_pixel_size(32)
-        status_row.append(self._meshbot_status_icon)
-
-        status_info = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=3)
-        self._meshbot_status_label = Gtk.Label(label="Checking...")
-        self._meshbot_status_label.set_xalign(0)
-        self._meshbot_status_label.add_css_class("heading")
-        status_info.append(self._meshbot_status_label)
-
-        self._meshbot_detail_label = Gtk.Label(label="")
-        self._meshbot_detail_label.set_xalign(0)
-        self._meshbot_detail_label.add_css_class("dim-label")
-        status_info.append(self._meshbot_detail_label)
-
-        status_row.append(status_info)
-
-        spacer = Gtk.Box()
-        spacer.set_hexpand(True)
-        status_row.append(spacer)
-
-        # Control buttons
-        btn_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-
-        self._start_btn = Gtk.Button(label="Start")
-        self._start_btn.add_css_class("suggested-action")
-        self._start_btn.connect("clicked", self._on_start_bot)
-        btn_box.append(self._start_btn)
-
-        self._stop_btn = Gtk.Button(label="Stop")
-        self._stop_btn.add_css_class("destructive-action")
-        self._stop_btn.connect("clicked", self._on_stop_bot)
-        btn_box.append(self._stop_btn)
-
-        install_btn = Gtk.Button(label="Install")
-        install_btn.connect("clicked", self._on_install_bot)
-        btn_box.append(install_btn)
-
-        status_row.append(btn_box)
-        status_box.append(status_row)
-
-        status_frame.set_child(status_box)
-        box.append(status_frame)
-
-        # Config section
-        config_frame = Gtk.Frame()
-        config_frame.set_label("Configuration")
-        config_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        config_box.set_margin_start(15)
-        config_box.set_margin_end(15)
-        config_box.set_margin_top(10)
-        config_box.set_margin_bottom(10)
-
-        # Path entry
-        path_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        path_row.append(Gtk.Label(label="Install Path:"))
-        self._path_entry = Gtk.Entry()
-        self._path_entry.set_text(self._settings.get("meshbot_path", "/opt/meshing-around"))
-        self._path_entry.set_hexpand(True)
-        path_row.append(self._path_entry)
-        config_box.append(path_row)
-
-        # Config file buttons
-        config_btn_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-
-        edit_config_btn = Gtk.Button(label="Edit Config")
-        edit_config_btn.connect("clicked", self._on_edit_config)
-        edit_config_btn.set_tooltip_text("Edit config.ini in text editor")
-        config_btn_row.append(edit_config_btn)
-
-        view_config_btn = Gtk.Button(label="View Config")
-        view_config_btn.connect("clicked", self._on_view_config)
-        view_config_btn.set_tooltip_text("View config.ini in log viewer below")
-        config_btn_row.append(view_config_btn)
-
-        open_folder_btn = Gtk.Button(label="Open Folder")
-        open_folder_btn.connect("clicked", self._on_open_meshbot_folder)
-        config_btn_row.append(open_folder_btn)
-
-        config_box.append(config_btn_row)
-
-        # Log/Debug buttons row
-        log_btn_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-
-        view_log_btn = Gtk.Button(label="View Log")
-        view_log_btn.connect("clicked", self._on_view_meshbot_log)
-        view_log_btn.set_tooltip_text("View MeshBot log file")
-        log_btn_row.append(view_log_btn)
-
-        tail_log_btn = Gtk.Button(label="Tail Log")
-        tail_log_btn.connect("clicked", self._on_tail_meshbot_log)
-        tail_log_btn.set_tooltip_text("Show last 50 lines of log")
-        log_btn_row.append(tail_log_btn)
-
-        journal_btn = Gtk.Button(label="Journal")
-        journal_btn.connect("clicked", self._on_view_journal)
-        journal_btn.set_tooltip_text("View systemd journal for mesh_bot")
-        log_btn_row.append(journal_btn)
-
-        config_box.append(log_btn_row)
-
-        config_frame.set_child(config_box)
-        box.append(config_frame)
-
-        # Connection Mode section - IMPORTANT for browser compatibility
-        conn_frame = Gtk.Frame()
-        conn_frame.set_label("Connection Mode")
-        conn_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        conn_box.set_margin_start(15)
-        conn_box.set_margin_end(15)
-        conn_box.set_margin_top(10)
-        conn_box.set_margin_bottom(10)
-
-        # Warning info
-        warn_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        warn_icon = Gtk.Image.new_from_icon_name("dialog-warning-symbolic")
-        warn_box.append(warn_icon)
-        warn_label = Gtk.Label(
-            label="Serial mode blocks browser access. Use TCP mode for shared access."
-        )
-        warn_label.set_wrap(True)
-        warn_label.set_xalign(0)
-        warn_label.add_css_class("dim-label")
-        warn_box.append(warn_label)
-        conn_box.append(warn_box)
-
-        # meshtasticd status
-        meshtasticd_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        meshtasticd_row.append(Gtk.Label(label="meshtasticd:"))
-        self._meshtasticd_status = Gtk.Label(label="Checking...")
-        self._meshtasticd_status.add_css_class("dim-label")
-        meshtasticd_row.append(self._meshtasticd_status)
-
-        spacer2 = Gtk.Box()
-        spacer2.set_hexpand(True)
-        meshtasticd_row.append(spacer2)
-
-        start_meshtasticd_btn = Gtk.Button(label="Start meshtasticd")
-        start_meshtasticd_btn.connect("clicked", self._on_start_meshtasticd)
-        meshtasticd_row.append(start_meshtasticd_btn)
-
-        conn_box.append(meshtasticd_row)
-
-        # Connection mode selector
-        mode_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        mode_row.append(Gtk.Label(label="MeshBot connects via:"))
-
-        self._conn_mode = Gtk.ComboBoxText()
-        self._conn_mode.append("serial", "Serial (exclusive - blocks browser)")
-        self._conn_mode.append("tcp", "TCP (shared - browser compatible)")
-        self._conn_mode.set_active_id("tcp")
-        self._conn_mode.set_tooltip_text("TCP mode requires meshtasticd running")
-        mode_row.append(self._conn_mode)
-
-        apply_mode_btn = Gtk.Button(label="Apply to Config")
-        apply_mode_btn.add_css_class("suggested-action")
-        apply_mode_btn.connect("clicked", self._on_apply_connection_mode)
-        apply_mode_btn.set_tooltip_text("Update mesh_bot config.ini with selected mode")
-        mode_row.append(apply_mode_btn)
-
-        conn_box.append(mode_row)
-
-        # Quick browser access
-        browser_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        browser_label = Gtk.Label(label="Quick access:")
-        browser_label.add_css_class("dim-label")
-        browser_row.append(browser_label)
-
-        open_browser_btn = Gtk.Button(label="Open Meshtastic Web")
-        open_browser_btn.connect("clicked", self._on_open_meshtastic_web)
-        open_browser_btn.set_tooltip_text("Open Meshtastic web interface (stops MeshBot if running in serial mode)")
-        browser_row.append(open_browser_btn)
-
-        conn_box.append(browser_row)
-
-        conn_frame.set_child(conn_box)
-        box.append(conn_frame)
-
-        # Features overview
-        features_frame = Gtk.Frame()
-        features_frame.set_label("MeshBot Features")
-        features_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=5)
-        features_box.set_margin_start(15)
-        features_box.set_margin_end(15)
-        features_box.set_margin_top(10)
-        features_box.set_margin_bottom(10)
-
-        features = [
-            ("BBS Messaging", "Store-and-forward bulletin board"),
-            ("Weather/Alerts", "NOAA, USGS, FEMA emergency data"),
-            ("Games", "DopeWars, BlackJack, Poker"),
-            ("Inventory/POS", "Point-of-sale with cart system"),
-        ]
-
-        for title, desc in features:
-            row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-            title_lbl = Gtk.Label(label=f"{title}:")
-            title_lbl.set_xalign(0)
-            title_lbl.add_css_class("heading")
-            title_lbl.set_width_chars(15)
-            row.append(title_lbl)
-
-            desc_lbl = Gtk.Label(label=desc)
-            desc_lbl.set_xalign(0)
-            desc_lbl.add_css_class("dim-label")
-            row.append(desc_lbl)
-            features_box.append(row)
-
-        features_frame.set_child(features_box)
-        box.append(features_frame)
-
-        scrolled.set_child(box)
-
-        # Tab label with icon
-        tab_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        tab_box.append(Gtk.Image.new_from_icon_name("mail-send-symbolic"))
-        tab_box.append(Gtk.Label(label="MeshBot"))
-
-        self._notebook.append_page(scrolled, tab_box)
-
-    # =========================================================================
-    # Node Map Tab
-    # =========================================================================
-
-    def _add_map_tab(self):
-        """Add Node Map tab"""
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=15)
-        box.set_margin_start(15)
-        box.set_margin_end(15)
-        box.set_margin_top(15)
-        box.set_margin_bottom(15)
-
-        # Map controls
-        controls_frame = Gtk.Frame()
-        controls_frame.set_label("Node Map Controls")
-        controls_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        controls_box.set_margin_start(15)
-        controls_box.set_margin_end(15)
-        controls_box.set_margin_top(10)
-        controls_box.set_margin_bottom(10)
-
-        btn_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-
-        refresh_map_btn = Gtk.Button(label="Refresh Nodes")
-        refresh_map_btn.add_css_class("suggested-action")
-        refresh_map_btn.connect("clicked", self._on_refresh_map)
-        btn_row.append(refresh_map_btn)
-
-        open_map_btn = Gtk.Button(label="Open Full Map")
-        open_map_btn.connect("clicked", self._on_open_full_map)
-        open_map_btn.set_tooltip_text("Open node map in browser")
-        btn_row.append(open_map_btn)
-
-        export_btn = Gtk.Button(label="Export GeoJSON")
-        export_btn.connect("clicked", self._on_export_geojson)
-        btn_row.append(export_btn)
-
-        controls_box.append(btn_row)
-        controls_frame.set_child(controls_box)
-        box.append(controls_frame)
-
-        # Node list
-        nodes_frame = Gtk.Frame()
-        nodes_frame.set_label("Discovered Nodes")
-        nodes_frame.set_vexpand(True)
-        nodes_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        nodes_box.set_margin_start(15)
-        nodes_box.set_margin_end(15)
-        nodes_box.set_margin_top(10)
-        nodes_box.set_margin_bottom(10)
-
-        # Node list store
-        self._node_store = Gtk.ListStore(str, str, str, str)  # ID, Name, Type, Last Seen
-
-        self._node_tree = Gtk.TreeView(model=self._node_store)
-        self._node_tree.set_headers_visible(True)
-
-        for i, title in enumerate(["Node ID", "Name", "Type", "Last Seen"]):
-            renderer = Gtk.CellRendererText()
-            column = Gtk.TreeViewColumn(title, renderer, text=i)
-            column.set_resizable(True)
-            column.set_min_width(80)
-            self._node_tree.append_column(column)
-
-        node_scroll = Gtk.ScrolledWindow()
-        node_scroll.set_policy(Gtk.PolicyType.AUTOMATIC, Gtk.PolicyType.AUTOMATIC)
-        node_scroll.set_min_content_height(150)
-        node_scroll.set_vexpand(True)
-        node_scroll.set_child(self._node_tree)
-
-        nodes_box.append(node_scroll)
-
-        # Stats row
-        self._node_stats_label = Gtk.Label(label="Nodes: 0 | Last update: Never")
-        self._node_stats_label.set_xalign(0)
-        self._node_stats_label.add_css_class("dim-label")
-        nodes_box.append(self._node_stats_label)
-
-        nodes_frame.set_child(nodes_box)
-        box.append(nodes_frame)
-
-        scrolled.set_child(box)
-
-        # Tab label
-        tab_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        tab_box.append(Gtk.Image.new_from_icon_name("mark-location-symbolic"))
-        tab_box.append(Gtk.Label(label="Node Map"))
-
-        self._notebook.append_page(scrolled, tab_box)
-
-    # =========================================================================
-    # Diagnostics Tab
-    # =========================================================================
-
-    def _add_diagnostics_tab(self):
-        """Add Diagnostics tab"""
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=15)
-        box.set_margin_start(15)
-        box.set_margin_end(15)
-        box.set_margin_top(15)
-        box.set_margin_bottom(15)
-
-        # Health status
-        health_frame = Gtk.Frame()
-        health_frame.set_label("System Health")
-        health_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        health_box.set_margin_start(15)
-        health_box.set_margin_end(15)
-        health_box.set_margin_top(10)
-        health_box.set_margin_bottom(10)
-
-        # Health cards grid
-        health_grid = Gtk.Grid()
-        health_grid.set_column_spacing(15)
-        health_grid.set_row_spacing(10)
-
-        self._health_cards = {}
-        checks = [
-            ("meshtastic", "Meshtastic", "network-wireless-symbolic"),
-            ("reticulum", "Reticulum", "network-transmit-receive-symbolic"),
-            ("meshbot", "MeshBot", "mail-send-symbolic"),
-            ("network", "Network", "network-wired-symbolic"),
-        ]
-
-        for i, (key, label, icon) in enumerate(checks):
-            card = self._create_health_card(label, icon)
-            health_grid.attach(card, i % 2, i // 2, 1, 1)
-            self._health_cards[key] = card
-
-        health_box.append(health_grid)
-
-        # Refresh button
-        refresh_health_btn = Gtk.Button(label="Run Health Check")
-        refresh_health_btn.connect("clicked", self._on_run_health_check)
-        health_box.append(refresh_health_btn)
-
-        health_frame.set_child(health_box)
-        box.append(health_frame)
-
-        # Quick tests
-        tests_frame = Gtk.Frame()
-        tests_frame.set_label("Network Tests")
-        tests_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        tests_box.set_margin_start(15)
-        tests_box.set_margin_end(15)
-        tests_box.set_margin_top(10)
-        tests_box.set_margin_bottom(10)
-
-        tests_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-
-        ping_btn = Gtk.Button(label="Ping Test")
-        ping_btn.connect("clicked", self._on_ping_test)
-        tests_row.append(ping_btn)
-
-        traceroute_btn = Gtk.Button(label="Traceroute")
-        traceroute_btn.connect("clicked", self._on_traceroute_test)
-        tests_row.append(traceroute_btn)
-
-        dns_btn = Gtk.Button(label="DNS Check")
-        dns_btn.connect("clicked", self._on_dns_test)
-        tests_row.append(dns_btn)
-
-        port_btn = Gtk.Button(label="Port Scan")
-        port_btn.connect("clicked", self._on_port_scan)
-        tests_row.append(port_btn)
-
-        tests_box.append(tests_row)
-        tests_frame.set_child(tests_box)
-        box.append(tests_frame)
-
-        # Report generation
-        report_frame = Gtk.Frame()
-        report_frame.set_label("Diagnostic Reports")
-        report_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        report_box.set_margin_start(15)
-        report_box.set_margin_end(15)
-        report_box.set_margin_top(10)
-        report_box.set_margin_bottom(10)
-
-        gen_report_btn = Gtk.Button(label="Generate Full Report")
-        gen_report_btn.add_css_class("suggested-action")
-        gen_report_btn.connect("clicked", self._on_generate_report)
-        report_box.append(gen_report_btn)
-
-        open_reports_btn = Gtk.Button(label="Open Reports Folder")
-        open_reports_btn.connect("clicked", self._on_open_reports_folder)
-        report_box.append(open_reports_btn)
-
-        report_frame.set_child(report_box)
-        box.append(report_frame)
-
-        scrolled.set_child(box)
-
-        # Tab label
-        tab_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        tab_box.append(Gtk.Image.new_from_icon_name("dialog-information-symbolic"))
-        tab_box.append(Gtk.Label(label="Diagnostics"))
-
-        self._notebook.append_page(scrolled, tab_box)
-
-    # =========================================================================
-    # Sensors Tab - Display telemetry from mesh nodes
-    # =========================================================================
-
-    def _add_sensors_tab(self):
-        """Add Sensors tab showing telemetry data from mesh nodes"""
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=15)
-        box.set_margin_start(15)
-        box.set_margin_end(15)
-        box.set_margin_top(15)
-        box.set_margin_bottom(15)
-
-        # Info header
-        info_label = Gtk.Label(
-            label="Telemetry data from Meshtastic nodes with sensors (BME280, PMSA003I, etc.)"
-        )
-        info_label.set_xalign(0)
-        info_label.add_css_class("dim-label")
-        info_label.set_wrap(True)
-        box.append(info_label)
-
-        # Environment Sensors Frame
-        env_frame = Gtk.Frame()
-        env_frame.set_label("Environment Sensors")
-        env_frame.set_vexpand(True)
-        env_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        env_box.set_margin_start(15)
-        env_box.set_margin_end(15)
-        env_box.set_margin_top(10)
-        env_box.set_margin_bottom(10)
-
-        # Environment sensor list: Node, Temp, Humidity, Pressure, Battery, Last Update
-        self._env_store = Gtk.ListStore(str, str, str, str, str, str)
-        self._env_tree = Gtk.TreeView(model=self._env_store)
-        self._env_tree.set_headers_visible(True)
-
-        env_columns = [
-            ("Node", 150),
-            ("Temperature", 100),
-            ("Humidity", 100),
-            ("Pressure", 100),
-            ("Battery", 80),
-            ("Last Update", 120),
-        ]
-
-        for i, (title, width) in enumerate(env_columns):
-            renderer = Gtk.CellRendererText()
-            column = Gtk.TreeViewColumn(title, renderer, text=i)
-            column.set_resizable(True)
-            column.set_min_width(width)
-            self._env_tree.append_column(column)
-
-        env_scroll = Gtk.ScrolledWindow()
-        env_scroll.set_policy(Gtk.PolicyType.AUTOMATIC, Gtk.PolicyType.AUTOMATIC)
-        env_scroll.set_min_content_height(120)
-        env_scroll.set_vexpand(True)
-        env_scroll.set_child(self._env_tree)
-        env_box.append(env_scroll)
-
-        env_frame.set_child(env_box)
-        box.append(env_frame)
-
-        # Air Quality Sensors Frame
-        aq_frame = Gtk.Frame()
-        aq_frame.set_label("Air Quality Sensors")
-        aq_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        aq_box.set_margin_start(15)
-        aq_box.set_margin_end(15)
-        aq_box.set_margin_top(10)
-        aq_box.set_margin_bottom(10)
-
-        # Air quality list: Node, PM2.5, PM10, CO2, IAQ, Last Update
-        self._aq_store = Gtk.ListStore(str, str, str, str, str, str)
-        self._aq_tree = Gtk.TreeView(model=self._aq_store)
-        self._aq_tree.set_headers_visible(True)
-
-        aq_columns = [
-            ("Node", 150),
-            ("PM2.5", 80),
-            ("PM10", 80),
-            ("CO2 (ppm)", 100),
-            ("IAQ", 80),
-            ("Last Update", 120),
-        ]
-
-        for i, (title, width) in enumerate(aq_columns):
-            renderer = Gtk.CellRendererText()
-            column = Gtk.TreeViewColumn(title, renderer, text=i)
-            column.set_resizable(True)
-            column.set_min_width(width)
-            self._aq_tree.append_column(column)
-
-        aq_scroll = Gtk.ScrolledWindow()
-        aq_scroll.set_policy(Gtk.PolicyType.AUTOMATIC, Gtk.PolicyType.AUTOMATIC)
-        aq_scroll.set_min_content_height(100)
-        aq_scroll.set_child(self._aq_tree)
-        aq_box.append(aq_scroll)
-
-        aq_frame.set_child(aq_box)
-        box.append(aq_frame)
-
-        # Detection Sensors Frame
-        det_frame = Gtk.Frame()
-        det_frame.set_label("Detection Sensors (Motion/Door Alerts)")
-        det_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        det_box.set_margin_start(15)
-        det_box.set_margin_end(15)
-        det_box.set_margin_top(10)
-        det_box.set_margin_bottom(10)
-
-        # Detection sensor list: Node, Sensor, State, Last Triggered, Count
-        self._det_store = Gtk.ListStore(str, str, str, str, str)
-        self._det_tree = Gtk.TreeView(model=self._det_store)
-        self._det_tree.set_headers_visible(True)
-
-        det_columns = [
-            ("Node", 150),
-            ("Sensor", 120),
-            ("State", 100),
-            ("Last Triggered", 150),
-            ("Count", 80),
-        ]
-
-        for i, (title, width) in enumerate(det_columns):
-            renderer = Gtk.CellRendererText()
-            column = Gtk.TreeViewColumn(title, renderer, text=i)
-            column.set_resizable(True)
-            column.set_min_width(width)
-            self._det_tree.append_column(column)
-
-        det_scroll = Gtk.ScrolledWindow()
-        det_scroll.set_policy(Gtk.PolicyType.AUTOMATIC, Gtk.PolicyType.AUTOMATIC)
-        det_scroll.set_min_content_height(80)
-        det_scroll.set_child(self._det_tree)
-        det_box.append(det_scroll)
-
-        det_frame.set_child(det_box)
-        box.append(det_frame)
-
-        # Controls row
-        controls = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-
-        refresh_btn = Gtk.Button(label="Refresh Sensors")
-        refresh_btn.connect("clicked", self._on_refresh_sensors)
-        controls.append(refresh_btn)
-
-        self._sensor_stats_label = Gtk.Label(label="No sensor data loaded")
-        self._sensor_stats_label.set_xalign(0)
-        self._sensor_stats_label.add_css_class("dim-label")
-        self._sensor_stats_label.set_hexpand(True)
-        controls.append(self._sensor_stats_label)
-
-        box.append(controls)
-
-        scrolled.set_child(box)
-
-        # Tab label
-        tab_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        tab_box.append(Gtk.Image.new_from_icon_name("weather-few-clouds-symbolic"))
-        tab_box.append(Gtk.Label(label="Sensors"))
-
-        self._notebook.append_page(scrolled, tab_box)
-
-    def _on_refresh_sensors(self, button):
-        """Refresh sensor data from mesh nodes"""
-        button.set_sensitive(False)
-        self._sensor_stats_label.set_label("Loading sensor data...")
-
-        def do_refresh():
-            env_nodes = []
-            aq_nodes = []
-            det_alerts = []
-
-            try:
-                # Try to get data from node tracker
-                from gateway.node_tracker import NodeTracker
-                tracker = NodeTracker.get_instance() if hasattr(NodeTracker, 'get_instance') else None
-
-                if tracker and hasattr(tracker, 'get_all_nodes'):
-                    for node in tracker.get_all_nodes():
-                        if not hasattr(node, 'telemetry') or not node.telemetry:
-                            continue
-
-                        t = node.telemetry
-                        node_name = node.name or node.id[:12]
-                        last_update = t.timestamp.strftime("%H:%M:%S") if t.timestamp else "N/A"
-
-                        # Environment sensors
-                        if t.temperature is not None or t.humidity is not None:
-                            temp = f"{t.temperature:.1f}°C" if t.temperature else "—"
-                            hum = f"{t.humidity:.0f}%" if t.humidity else "—"
-                            pres = f"{t.pressure:.0f} hPa" if t.pressure else "—"
-                            batt = f"{t.battery_level}%" if t.battery_level else "—"
-                            env_nodes.append((node_name, temp, hum, pres, batt, last_update))
-
-                        # Air quality sensors
-                        if t.air_quality and t.air_quality.has_data():
-                            aq = t.air_quality
-                            pm25 = str(aq.pm25_standard) if aq.pm25_standard else "—"
-                            pm10 = str(aq.pm100_standard) if aq.pm100_standard else "—"
-                            co2 = str(aq.co2) if aq.co2 else "—"
-                            iaq = str(aq.iaq) if aq.iaq else "—"
-                            aq_nodes.append((node_name, pm25, pm10, co2, iaq, last_update))
-
-                        # Detection sensors
-                        for sensor in t.detection_sensors:
-                            state = "TRIGGERED" if sensor.triggered else "Clear"
-                            last_trig = sensor.last_triggered.strftime("%H:%M:%S") if sensor.last_triggered else "Never"
-                            det_alerts.append((node_name, sensor.name, state, last_trig, str(sensor.trigger_count)))
-
-            except Exception as e:
-                logger.error(f"Error loading sensor data: {e}")
-                # Add placeholder if no data
-                if not env_nodes:
-                    env_nodes.append(("No nodes", "—", "—", "—", "—", "Connect to mesh"))
-
-            # Update UI
-            GLib.idle_add(self._update_sensor_display, env_nodes, aq_nodes, det_alerts)
-            GLib.idle_add(button.set_sensitive, True)
-
-        threading.Thread(target=do_refresh, daemon=True).start()
-
-    def _update_sensor_display(self, env_nodes, aq_nodes, det_alerts):
-        """Update sensor display lists (called on main thread)"""
-        # Update environment sensors
-        self._env_store.clear()
-        for row in env_nodes:
-            self._env_store.append(row)
-
-        # Update air quality sensors
-        self._aq_store.clear()
-        for row in aq_nodes:
-            self._aq_store.append(row)
-
-        # Update detection sensors
-        self._det_store.clear()
-        for row in det_alerts:
-            self._det_store.append(row)
-
-        # Update stats
-        total = len(env_nodes) + len(aq_nodes) + len(det_alerts)
-        self._sensor_stats_label.set_label(
-            f"Environment: {len(env_nodes)} | Air Quality: {len(aq_nodes)} | Detection: {len(det_alerts)}"
-        )
-        return False
-
-    def _create_health_card(self, label: str, icon_name: str) -> Gtk.Box:
-        """Create a health status card"""
-        card = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        card.add_css_class("card")
-        card.set_margin_start(10)
-        card.set_margin_end(10)
-        card.set_margin_top(5)
-        card.set_margin_bottom(5)
-
-        icon = Gtk.Image.new_from_icon_name(icon_name)
-        icon.set_pixel_size(24)
-        card.append(icon)
-
-        info = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        name = Gtk.Label(label=label)
-        name.set_xalign(0)
-        name.add_css_class("heading")
-        info.append(name)
-
-        status = Gtk.Label(label="Unknown")
-        status.set_xalign(0)
-        status.add_css_class("dim-label")
-        status.set_name("status")
-        info.append(status)
-
-        card.append(info)
-
-        # Store reference to status label
-        card._status_label = status
-        return card
-
-    # =========================================================================
-    # Event Handlers
+    # Status Checking (Orchestrates tab checks)
     # =========================================================================
 
     def _check_all_status(self):
@@ -1003,1073 +284,9 @@ class MeshToolsPanel(Gtk.Box):
         self._update_health_cards()
         return False  # Don't repeat
 
-    def _check_meshtasticd_status(self):
-        """Check if meshtasticd is running"""
-        def check():
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'meshtasticd'],
-                    capture_output=True, text=True, timeout=5
-                )
-                running = result.returncode == 0 and result.stdout.strip()
-                GLib.idle_add(self._update_meshtasticd_status, running)
-            except Exception:
-                GLib.idle_add(self._update_meshtasticd_status, False)
-
-        threading.Thread(target=check, daemon=True).start()
-
-    def _update_meshtasticd_status(self, running: bool):
-        """Update meshtasticd status display"""
-        # Guard: Panel might have been destroyed
-        if not hasattr(self, '_meshtasticd_status') or self._meshtasticd_status is None:
-            return False
-        try:
-            if running:
-                self._meshtasticd_status.set_label("Running (TCP available)")
-                self._meshtasticd_status.remove_css_class("error")
-                self._meshtasticd_status.add_css_class("success")
-            else:
-                self._meshtasticd_status.set_label("Not Running")
-                self._meshtasticd_status.remove_css_class("success")
-                self._meshtasticd_status.add_css_class("error")
-        except Exception:
-            pass
-        return False
-
-    def _on_start_meshtasticd(self, button):
-        """Start meshtasticd service"""
-        self._log_message("Starting meshtasticd...")
-
-        def do_start():
-            try:
-                # Try systemctl first
-                result = subprocess.run(
-                    ['sudo', 'systemctl', 'start', 'meshtasticd'],
-                    capture_output=True, text=True, timeout=30
-                )
-                if result.returncode == 0:
-                    GLib.idle_add(self._log_message, "meshtasticd started via systemctl")
-                else:
-                    # Try direct start
-                    subprocess.Popen(
-                        ['sudo', 'meshtasticd'],
-                        stdout=subprocess.DEVNULL,
-                        stderr=subprocess.DEVNULL,
-                        start_new_session=True
-                    )
-                    GLib.idle_add(self._log_message, "meshtasticd started directly")
-
-                import time
-                time.sleep(2)
-                GLib.idle_add(self._check_meshtasticd_status)
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Failed to start meshtasticd: {e}")
-
-        threading.Thread(target=do_start, daemon=True).start()
-
-    def _on_apply_connection_mode(self, button):
-        """Apply connection mode to mesh_bot config"""
-        mode = self._conn_mode.get_active_id()
-        meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-        config_path = Path(meshbot_path) / "config.ini"
-
-        if not config_path.exists():
-            self._log_message(f"Config not found: {config_path}")
-            return
-
-        self._log_message(f"Applying {mode} connection mode...")
-
-        def do_apply():
-            try:
-                import configparser
-                config = configparser.ConfigParser()
-                config.read(str(config_path))
-
-                # Update interface settings
-                if 'interface' not in config:
-                    config['interface'] = {}
-
-                if mode == 'tcp':
-                    config['interface']['type'] = 'tcp'
-                    config['interface']['hostname'] = 'localhost'
-                    config['interface']['port'] = '4403'
-                    # Remove serial settings
-                    if 'port' in config['interface'] and config['interface']['port'].startswith('/dev'):
-                        del config['interface']['port']
-                else:  # serial
-                    config['interface']['type'] = 'serial'
-                    # Default to common serial ports
-                    if 'port' not in config['interface'] or not config['interface']['port'].startswith('/dev'):
-                        config['interface']['port'] = '/dev/ttyUSB0'
-
-                # Write config
-                with open(str(config_path), 'w') as f:
-                    config.write(f)
-
-                GLib.idle_add(self._log_message, f"Config updated to {mode} mode")
-                GLib.idle_add(self._log_message, "Restart MeshBot for changes to take effect")
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Failed to update config: {e}")
-
-        threading.Thread(target=do_apply, daemon=True).start()
-
-    def _on_open_meshtastic_web(self, button):
-        """Open Meshtastic web interface"""
-        # Check if mesh_bot is running in serial mode - offer to stop it
-        def check_and_open():
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'mesh_bot.py'],
-                    capture_output=True, text=True, timeout=5
-                )
-                meshbot_running = result.returncode == 0 and result.stdout.strip()
-
-                if meshbot_running:
-                    GLib.idle_add(self._log_message, "MeshBot is running - may conflict with browser")
-                    GLib.idle_add(self._log_message, "Consider using TCP mode or stopping MeshBot")
-
-                # Open web interface anyway (meshtasticd uses port 9443 with HTTPS)
-                GLib.idle_add(self._open_url, "https://localhost:9443")
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Error: {e}")
-                GLib.idle_add(self._open_url, "https://localhost:9443")
-
-        threading.Thread(target=check_and_open, daemon=True).start()
-
-    def _check_meshbot_status(self):
-        """Check MeshBot installation and running status"""
-        def check():
-            status = {"installed": False, "running": False, "path": None}
-
-            meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-
-            if Path(meshbot_path).exists() and (Path(meshbot_path) / "mesh_bot.py").exists():
-                status["installed"] = True
-                status["path"] = meshbot_path
-
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'mesh_bot.py'],
-                    capture_output=True, text=True, timeout=5
-                )
-                if result.returncode == 0 and result.stdout.strip():
-                    status["running"] = True
-            except Exception:
-                pass
-
-            GLib.idle_add(self._update_meshbot_status, status)
-
-        threading.Thread(target=check, daemon=True).start()
-
-    def _update_meshbot_status(self, status):
-        """Update MeshBot status display"""
-        if status["running"]:
-            self._meshbot_status_icon.set_from_icon_name("emblem-default-symbolic")
-            self._meshbot_status_label.set_label("MeshBot Running")
-            self._meshbot_detail_label.set_label(f"Path: {status['path']}")
-            self._start_btn.set_sensitive(False)
-            self._stop_btn.set_sensitive(True)
-        elif status["installed"]:
-            self._meshbot_status_icon.set_from_icon_name("media-playback-stop-symbolic")
-            self._meshbot_status_label.set_label("MeshBot Stopped")
-            self._meshbot_detail_label.set_label(f"Path: {status['path']}")
-            self._start_btn.set_sensitive(True)
-            self._stop_btn.set_sensitive(False)
-        else:
-            self._meshbot_status_icon.set_from_icon_name("dialog-warning-symbolic")
-            self._meshbot_status_label.set_label("MeshBot Not Installed")
-            self._meshbot_detail_label.set_label("Click Install to set up")
-            self._start_btn.set_sensitive(False)
-            self._stop_btn.set_sensitive(False)
-
-    def _update_health_cards(self):
-        """Update health status cards"""
-        for key, card in self._health_cards.items():
-            # Default to unknown
-            card._status_label.set_label("Checking...")
-
-        # Check each service
-        def check_health():
-            results = {}
-
-            # Check Meshtastic
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'meshtasticd'],
-                    capture_output=True, timeout=5
-                )
-                results["meshtastic"] = "Running" if result.returncode == 0 else "Not Running"
-            except Exception:
-                results["meshtastic"] = "Unknown"
-
-            # Check Reticulum
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'rnsd'],
-                    capture_output=True, timeout=5
-                )
-                results["reticulum"] = "Running" if result.returncode == 0 else "Not Running"
-            except Exception:
-                results["reticulum"] = "Unknown"
-
-            # Check MeshBot
-            try:
-                result = subprocess.run(
-                    ['pgrep', '-f', 'mesh_bot.py'],
-                    capture_output=True, timeout=5
-                )
-                results["meshbot"] = "Running" if result.returncode == 0 else "Not Running"
-            except Exception:
-                results["meshbot"] = "Unknown"
-
-            # Check Network
-            try:
-                result = subprocess.run(
-                    ['ping', '-c', '1', '-W', '2', '8.8.8.8'],
-                    capture_output=True, timeout=5
-                )
-                results["network"] = "Connected" if result.returncode == 0 else "No Internet"
-            except Exception:
-                results["network"] = "Unknown"
-
-            GLib.idle_add(self._apply_health_results, results)
-
-        threading.Thread(target=check_health, daemon=True).start()
-
-    def _apply_health_results(self, results):
-        """Apply health check results to cards"""
-        for key, status in results.items():
-            if key in self._health_cards:
-                card = self._health_cards[key]
-                card._status_label.set_label(status)
-
-                # Color code
-                if "Running" in status or "Connected" in status:
-                    card._status_label.remove_css_class("error")
-                    card._status_label.add_css_class("success")
-                elif "Not" in status or "No " in status:
-                    card._status_label.remove_css_class("success")
-                    card._status_label.add_css_class("error")
-
-    def _on_start_bot(self, button):
-        """Start MeshBot - use systemd service if available, otherwise direct execution"""
-        meshbot_path = self._path_entry.get_text().strip()
-        script_path = Path(meshbot_path) / "mesh_bot.py"
-
-        if not script_path.exists():
-            self._log_message("Error: mesh_bot.py not found")
-            return
-
-        # Check for config.ini
-        config_path = Path(meshbot_path) / "config.ini"
-        if not config_path.exists():
-            self._log_message("Warning: config.ini not found - creating from template...")
-            template = Path(meshbot_path) / "config.template"
-            if template.exists():
-                real_user = os.environ.get('SUDO_USER', os.environ.get('USER', 'pi'))
-                subprocess.run(['sudo', 'cp', str(template), str(config_path)], timeout=10)
-                subprocess.run(['sudo', 'chown', f'{real_user}:{real_user}', str(config_path)], timeout=10)
-            else:
-                self._log_message("Error: No config.ini or config.template found")
-                return
-
-        self._log_message("Starting MeshBot...")
-
-        def do_start():
-            try:
-                # Check if systemd service exists
-                service_check = subprocess.run(
-                    ['systemctl', 'cat', 'mesh_bot.service'],
-                    capture_output=True, text=True, timeout=5
-                )
-
-                if service_check.returncode == 0:
-                    # Use systemd service
-                    GLib.idle_add(self._log_message, "Using systemd service...")
-                    result = subprocess.run(
-                        ['sudo', 'systemctl', 'start', 'mesh_bot.service'],
-                        capture_output=True, text=True, timeout=30
-                    )
-                    if result.returncode == 0:
-                        GLib.idle_add(self._log_message, "MeshBot service started")
-                        GLib.idle_add(self._check_meshbot_status)
-                    else:
-                        GLib.idle_add(self._log_message, f"Service start failed: {result.stderr}")
-                    return
-
-                # Fall back to direct execution
-                # Check for virtual environment
-                venv_path = Path(meshbot_path) / "venv"
-                venv_python = venv_path / "bin" / "python3"
-
-                if venv_python.exists():
-                    # Use venv python
-                    cmd = [str(venv_python), str(script_path)]
-                    GLib.idle_add(self._log_message, "Using virtual environment")
-                else:
-                    # Run directly with system python3
-                    cmd = ['python3', str(script_path)]
-                    GLib.idle_add(self._log_message, "Using system Python (no venv found)")
-
-                GLib.idle_add(self._log_message, f"Running: {' '.join(cmd)}")
-
-                process = subprocess.Popen(
-                    cmd,
-                    cwd=meshbot_path,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.STDOUT,
-                    text=True,
-                    start_new_session=True
-                )
-
-                import time
-                time.sleep(3)  # Give it a bit more time
-
-                if process.poll() is None:
-                    GLib.idle_add(self._log_message, "MeshBot started successfully")
-                    GLib.idle_add(self._check_meshbot_status)
-                else:
-                    # Process exited - capture output to show why
-                    exit_code = process.returncode
-                    output = process.stdout.read() if process.stdout else ""
-                    GLib.idle_add(self._log_message, f"MeshBot exited with code {exit_code}")
-                    if output:
-                        # Show first few lines of error
-                        lines = output.strip().split('\n')[:10]
-                        for line in lines:
-                            GLib.idle_add(self._log_message, f"  {line}")
-                    # Suggest venv setup if missing
-                    if not venv_python.exists():
-                        GLib.idle_add(self._log_message, "Tip: Create venv with:")
-                        GLib.idle_add(self._log_message, f"  cd {meshbot_path} && python3 -m venv venv")
-                        GLib.idle_add(self._log_message, "  venv/bin/pip install -r requirements.txt")
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Start error: {e}")
-
-        threading.Thread(target=do_start, daemon=True).start()
-
-    def _on_stop_bot(self, button):
-        """Stop MeshBot - use systemd service if available"""
-        self._log_message("Stopping MeshBot...")
-
-        def do_stop():
-            try:
-                # Check if systemd service exists and is running
-                service_check = subprocess.run(
-                    ['systemctl', 'is-active', 'mesh_bot.service'],
-                    capture_output=True, text=True, timeout=5
-                )
-
-                if service_check.stdout.strip() == 'active':
-                    # Stop via systemd
-                    GLib.idle_add(self._log_message, "Stopping systemd service...")
-                    result = subprocess.run(
-                        ['sudo', 'systemctl', 'stop', 'mesh_bot.service'],
-                        capture_output=True, text=True, timeout=30
-                    )
-                    if result.returncode == 0:
-                        GLib.idle_add(self._log_message, "MeshBot service stopped")
-                    else:
-                        GLib.idle_add(self._log_message, f"Service stop failed: {result.stderr}")
-                else:
-                    # Kill any running process
-                    subprocess.run(['pkill', '-f', 'mesh_bot.py'], timeout=10)
-                    GLib.idle_add(self._log_message, "MeshBot stopped")
-
-                import time
-                time.sleep(1)
-                GLib.idle_add(self._check_meshbot_status)
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Stop error: {e}")
-
-        threading.Thread(target=do_stop, daemon=True).start()
-
-    def _on_install_bot(self, button):
-        """Install MeshBot"""
-        install_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-        self._log_message(f"Installing MeshBot to {install_path}...")
-
-        def do_install():
-            try:
-                install_dir = Path(install_path)
-                real_user = os.environ.get('SUDO_USER', os.environ.get('USER', 'pi'))
-
-                if install_dir.exists():
-                    GLib.idle_add(self._log_message, "Updating existing installation...")
-                    result = subprocess.run(
-                        ['git', '-C', str(install_dir), 'pull'],
-                        capture_output=True, text=True, timeout=120
-                    )
-                else:
-                    GLib.idle_add(self._log_message, "Cloning repository...")
-                    result = subprocess.run(
-                        ['sudo', 'git', 'clone',
-                         'https://github.com/SpudGunMan/meshing-around.git',
-                         str(install_dir)],
-                        capture_output=True, text=True, timeout=300
-                    )
-
-                if result.returncode != 0:
-                    GLib.idle_add(self._log_message, f"Install failed: {result.stderr}")
-                    return
-
-                # Set ownership to real user
-                GLib.idle_add(self._log_message, "Setting permissions...")
-                subprocess.run(['sudo', 'chown', '-R', f'{real_user}:{real_user}', str(install_dir)], timeout=60)
-
-                # Set up virtual environment if not exists
-                venv_path = install_dir / "venv"
-                if not venv_path.exists():
-                    GLib.idle_add(self._log_message, "Creating virtual environment...")
-                    venv_result = subprocess.run(
-                        ['python3', '-m', 'venv', str(venv_path)],
-                        capture_output=True, text=True, timeout=120,
-                        cwd=str(install_dir)
-                    )
-                    if venv_result.returncode != 0:
-                        GLib.idle_add(self._log_message, f"Venv creation failed: {venv_result.stderr}")
-                        GLib.idle_add(self._log_message, "You can create it manually later")
-                    else:
-                        # Install requirements
-                        req_file = install_dir / "requirements.txt"
-                        if req_file.exists():
-                            GLib.idle_add(self._log_message, "Installing dependencies (this may take a while)...")
-                            pip_path = venv_path / "bin" / "pip"
-                            pip_result = subprocess.run(
-                                [str(pip_path), 'install', '-r', str(req_file)],
-                                capture_output=True, text=True, timeout=600,
-                                cwd=str(install_dir)
-                            )
-                            if pip_result.returncode == 0:
-                                GLib.idle_add(self._log_message, "Dependencies installed!")
-                            else:
-                                GLib.idle_add(self._log_message, f"Pip install warning: {pip_result.stderr[:200]}")
-
-                GLib.idle_add(self._log_message, "Installation complete!")
-                GLib.idle_add(self._check_meshbot_status)
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Install error: {e}")
-
-        threading.Thread(target=do_install, daemon=True).start()
-
-    def _on_edit_config(self, button):
-        """Open config file in editor"""
-        meshbot_path = self._path_entry.get_text().strip()
-        config_path = Path(meshbot_path) / "config.ini"
-
-        if not config_path.exists():
-            template = Path(meshbot_path) / "config.template"
-            if template.exists():
-                subprocess.run(['sudo', 'cp', str(template), str(config_path)], timeout=10)
-            else:
-                self._log_message("No config file found")
-                return
-
-        # Ensure config file is writable by real user
-        real_user = os.environ.get('SUDO_USER', os.environ.get('USER', 'pi'))
-        subprocess.run(['sudo', 'chown', f'{real_user}:{real_user}', str(config_path)], timeout=10)
-        subprocess.run(['sudo', 'chmod', '644', str(config_path)], timeout=10)
-
-        # Open in default editor
-        try:
-            subprocess.Popen(
-                ['sudo', '-u', real_user, 'xdg-open', str(config_path)],
-                start_new_session=True
-            )
-            self._log_message(f"Opening {config_path} in editor")
-        except Exception as e:
-            self._log_message(f"Error opening editor: {e}")
-
-    def _on_view_config(self, button):
-        """View config in log viewer"""
-        meshbot_path = self._path_entry.get_text().strip()
-        config_path = Path(meshbot_path) / "config.ini"
-
-        if config_path.exists():
-            content = config_path.read_text()
-            self._set_log_text(f"=== {config_path} ===\n\n{content}")
-        else:
-            self._log_message("Config file not found")
-
-    def _on_open_meshbot_folder(self, button):
-        """Open MeshBot folder"""
-        meshbot_path = self._path_entry.get_text().strip()
-        self._open_folder(meshbot_path)
-
-    def _on_view_meshbot_log(self, button):
-        """View MeshBot log file"""
-        meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-        log_files = [
-            # Current meshing-around log locations
-            Path(meshbot_path) / "logs" / "meshbot.log",
-            Path(meshbot_path) / "logs" / "messages.log",
-            # Legacy/alternative locations
-            Path(meshbot_path) / "prior_log.txt",
-            Path(meshbot_path) / "prior_debug_log.txt",
-        ]
-
-        for log_path in log_files:
-            if log_path.exists():
-                try:
-                    content = log_path.read_text()
-                    self._set_log_text(f"=== {log_path} ===\n\n{content[-10000:]}")  # Last 10k chars
-                    self._log_message(f"Loaded log from {log_path}")
-                    return
-                except Exception as e:
-                    self._log_message(f"Error reading {log_path}: {e}")
-
-        self._log_message("No log files found")
-        self._log_message(f"Checked: {meshbot_path}/logs/meshbot.log")
-        self._log_message("Logs are created after MeshBot runs with logging enabled")
-
-    def _on_tail_meshbot_log(self, button):
-        """Show last 50 lines of log"""
-        meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-        log_files = [
-            # Current meshing-around log locations
-            Path(meshbot_path) / "logs" / "meshbot.log",
-            Path(meshbot_path) / "logs" / "messages.log",
-            # Legacy/alternative locations
-            Path(meshbot_path) / "prior_log.txt",
-            Path(meshbot_path) / "prior_debug_log.txt",
-        ]
-
-        for log_path in log_files:
-            if log_path.exists():
-                try:
-                    lines = log_path.read_text().strip().split('\n')
-                    tail = '\n'.join(lines[-50:])
-                    self._set_log_text(f"=== Last 50 lines of {log_path.name} ===\n\n{tail}")
-                    return
-                except Exception as e:
-                    self._log_message(f"Error reading {log_path}: {e}")
-
-        self._log_message("No log files found")
-        self._log_message("Enable logging in config.ini: syslog_to_file = True")
-
-    def _on_view_journal(self, button):
-        """View systemd journal for mesh_bot service"""
-        self._log_message("Fetching journal entries...")
-
-        def fetch_journal():
-            try:
-                # First try the mesh_bot.service
-                result = subprocess.run(
-                    ['journalctl', '--no-pager', '-n', '100', '-u', 'mesh_bot.service'],
-                    capture_output=True, text=True, timeout=30
-                )
-
-                if result.stdout.strip() and '-- No entries --' not in result.stdout:
-                    GLib.idle_add(self._set_log_text, f"=== Journal (mesh_bot.service) ===\n\n{result.stdout}")
-                    return
-
-                # Try grep pattern
-                result2 = subprocess.run(
-                    ['journalctl', '--no-pager', '-n', '100', '-g', 'mesh_bot'],
-                    capture_output=True, text=True, timeout=30
-                )
-
-                if result2.stdout.strip() and '-- No entries --' not in result2.stdout:
-                    GLib.idle_add(self._set_log_text, f"=== Journal (mesh_bot) ===\n\n{result2.stdout}")
-                else:
-                    GLib.idle_add(self._log_message, "No journal entries found")
-                    GLib.idle_add(self._log_message, "Use View Log or Tail Log to see file-based logs")
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Journal error: {e}")
-
-        threading.Thread(target=fetch_journal, daemon=True).start()
-
-    def _on_refresh_map(self, button):
-        """Refresh node list from meshtasticd or mesh_bot data"""
-        self._log_message("Refreshing node list...")
-
-        def fetch_nodes():
-            import socket
-            nodes = []
-
-            # Method 1: Check if meshtasticd TCP port is available first
-            tcp_available = False
-            sock = None
-            try:
-                sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-                sock.settimeout(2)
-                result = sock.connect_ex(('localhost', 4403))
-                tcp_available = (result == 0)
-            except Exception:
-                pass
-            finally:
-                if sock:
-                    try:
-                        sock.close()
-                    except Exception:  # Ignore errors during cleanup
-                        pass
-
-            if tcp_available:
-                # Acquire global lock - meshtasticd only supports one TCP connection
-                lock_acquired = False
-                if HAS_MESHTASTIC_LOCK and MESHTASTIC_CONNECTION_LOCK:
-                    lock_acquired = MESHTASTIC_CONNECTION_LOCK.acquire(timeout=5.0)
-                    if not lock_acquired:
-                        GLib.idle_add(self._log_message, "Could not acquire connection lock (another operation in progress)")
-                    else:
-                        wait_for_cooldown()
-                else:
-                    lock_acquired = True  # No lock available, proceed anyway
-
-                if lock_acquired:
-                    interface = None
-                    try:
-                        import meshtastic.tcp_interface
-                        import time
-                        GLib.idle_add(self._log_message, "Connecting to meshtasticd via TCP...")
-
-                        interface = meshtastic.tcp_interface.TCPInterface('localhost', 4403)
-
-                        # Wait briefly for node info to populate
-                        time.sleep(2)
-
-                        # Get nodes from interface
-                        if hasattr(interface, 'nodes') and interface.nodes:
-                            for node_id, node in interface.nodes.items():
-                                user = node.get('user', {})
-                                name = user.get('longName', user.get('shortName', 'Unknown'))
-                                hw_model = user.get('hwModel', 'Unknown')
-
-                                last_heard = node.get('lastHeard', 0)
-                                if last_heard:
-                                    last_seen = datetime.fromtimestamp(last_heard).strftime('%H:%M:%S')
-                                else:
-                                    last_seen = 'Never'
-
-                                nodes.append((str(node_id), name, hw_model, last_seen))
-
-                        if nodes:
-                            GLib.idle_add(self._log_message, f"Found {len(nodes)} nodes via meshtastic TCP")
-
-                    except ImportError:
-                        GLib.idle_add(self._log_message, "Meshtastic Python library not installed")
-                        GLib.idle_add(self._log_message, "Install with: pip install meshtastic")
-                    except (SystemExit, KeyboardInterrupt, GeneratorExit):
-                        raise
-                    except BaseException as e:
-                        # Catch pyo3 PanicException and other crashes
-                        GLib.idle_add(self._log_message, f"Meshtastic error: {e}")
-                    finally:
-                        # Always close interface and release lock
-                        if interface:
-                            if HAS_MESHTASTIC_LOCK:
-                                safe_close_interface(interface)
-                            else:
-                                try:
-                                    interface.close()
-                                except Exception:
-                                    pass
-                        if HAS_MESHTASTIC_LOCK and MESHTASTIC_CONNECTION_LOCK:
-                            try:
-                                MESHTASTIC_CONNECTION_LOCK.release()
-                            except RuntimeError:
-                                pass
-            else:
-                GLib.idle_add(self._log_message, "meshtasticd TCP port 4403 not available")
-
-            # Method 2: Try reading mesh_bot data files
-            if not nodes:
-                meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-                data_files = [
-                    Path(meshbot_path) / "prior_node_db.json",
-                    Path(meshbot_path) / "node_db.json",
-                    Path(meshbot_path) / "data" / "nodes.json",
-                ]
-
-                found_file = False
-                for data_file in data_files:
-                    if data_file.exists():
-                        found_file = True
-                        try:
-                            with open(data_file) as f:
-                                data = json.load(f)
-
-                            if isinstance(data, dict):
-                                for node_id, node in data.items():
-                                    if isinstance(node, dict):
-                                        name = node.get('longName', node.get('shortName', node.get('name', 'Unknown')))
-                                        hw = node.get('hwModel', node.get('type', 'Unknown'))
-                                        last = node.get('lastHeard', node.get('last_seen', ''))
-                                        if isinstance(last, (int, float)) and last > 0:
-                                            last = datetime.fromtimestamp(last).strftime('%H:%M:%S')
-                                        nodes.append((str(node_id), name, hw, str(last) if last else 'Unknown'))
-
-                            if nodes:
-                                GLib.idle_add(self._log_message, f"Loaded {len(nodes)} nodes from {data_file.name}")
-                            break
-                        except Exception as e:
-                            GLib.idle_add(self._log_message, f"Error reading {data_file.name}: {e}")
-
-                if not found_file:
-                    GLib.idle_add(self._log_message, "No mesh_bot node database files found")
-
-            # Show help if no nodes found
-            if not nodes:
-                GLib.idle_add(self._log_message, "")
-                GLib.idle_add(self._log_message, "=== No nodes found ===")
-                GLib.idle_add(self._log_message, "To get node data:")
-                GLib.idle_add(self._log_message, "  1. Start meshtasticd (systemctl start meshtasticd)")
-                GLib.idle_add(self._log_message, "  2. Install meshtastic: pip install meshtastic")
-                GLib.idle_add(self._log_message, "  3. Or run MeshBot to create node database")
-
-            # Update UI
-            GLib.idle_add(self._update_node_list, nodes)
-
-        threading.Thread(target=fetch_nodes, daemon=True).start()
-
-    def _update_node_list(self, nodes: list):
-        """Update the node list store with new data"""
-        self._node_store.clear()
-
-        for node_id, name, hw_model, last_seen in nodes:
-            self._node_store.append([node_id, name, hw_model, last_seen])
-
-        count = len(nodes)
-        self._node_stats_label.set_label(f"Nodes: {count} | Last update: {datetime.now().strftime('%H:%M:%S')}")
-
-    def _on_open_full_map(self, button):
-        """Open map view - check which ports are available first"""
-        import socket
-
-        # Map URLs to check with their ports (9443 uses HTTPS)
-        map_options = [
-            ("https://localhost:9443", 9443, "Meshtastic Web"),  # meshtasticd (HTTPS)
-            ("http://localhost:8080", 8080, "Meshtastic Alt"),
-            ("http://localhost:5000", 5000, "Flask/MeshBot"),
-            ("http://localhost:8000", 8000, "MeshMap"),
-        ]
-
-        self._log_message("Checking available map servers...")
-        real_user = os.environ.get('SUDO_USER', os.environ.get('USER', 'pi'))
-
-        # Check which ports are actually open
-        for url, port, name in map_options:
-            sock = None
-            try:
-                sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-                sock.settimeout(1)
-                result = sock.connect_ex(('localhost', port))
-
-                if result == 0:
-                    self._log_message(f"Found {name} on port {port}")
-                    subprocess.Popen(
-                        ['sudo', '-u', real_user, 'xdg-open', url],
-                        stdout=subprocess.DEVNULL,
-                        stderr=subprocess.DEVNULL,
-                        start_new_session=True
-                    )
-                    self._log_message(f"Opening {url}")
-                    return
-            except Exception:
-                continue
-            finally:
-                if sock:
-                    try:
-                        sock.close()
-                    except Exception:
-                        pass
-
-        self._log_message("No map server found on common ports (9443, 8080, 5000, 8000)")
-        self._log_message("Make sure meshtasticd or meshbot web is running")
-
-    def _on_export_geojson(self, button):
-        """Export nodes as GeoJSON"""
-        self._log_message("Exporting GeoJSON...")
-
-        def do_export():
-            try:
-                # Collect nodes from store
-                nodes = []
-                model = self._node_store
-                iter = model.get_iter_first()
-                while iter:
-                    node_id = model.get_value(iter, 0)
-                    name = model.get_value(iter, 1)
-                    hw_model = model.get_value(iter, 2)
-                    last_seen = model.get_value(iter, 3)
-                    nodes.append({
-                        'id': node_id,
-                        'name': name,
-                        'hw_model': hw_model,
-                        'last_seen': last_seen
-                    })
-                    iter = model.iter_next(iter)
-
-                if not nodes:
-                    GLib.idle_add(self._log_message, "No nodes to export. Refresh node list first.")
-                    return
-
-                # Try to get coordinates from meshtasticd or mesh_bot data
-                meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-                node_coords = {}
-
-                # Try reading coordinate data from mesh_bot files
-                data_files = [
-                    Path(meshbot_path) / "prior_node_db.json",
-                    Path(meshbot_path) / "node_db.json",
-                ]
-                for data_file in data_files:
-                    if data_file.exists():
-                        try:
-                            with open(data_file) as f:
-                                data = json.load(f)
-                            for nid, node in data.items():
-                                if isinstance(node, dict):
-                                    pos = node.get('position', {})
-                                    lat = pos.get('latitude') or pos.get('lat')
-                                    lon = pos.get('longitude') or pos.get('lon')
-                                    if lat and lon:
-                                        node_coords[str(nid)] = (lon, lat)
-                            break
-                        except Exception:
-                            pass
-
-                # Build GeoJSON
-                features = []
-                for node in nodes:
-                    coords = node_coords.get(node['id'])
-                    if coords:
-                        feature = {
-                            "type": "Feature",
-                            "geometry": {
-                                "type": "Point",
-                                "coordinates": list(coords)
-                            },
-                            "properties": {
-                                "id": node['id'],
-                                "name": node['name'],
-                                "hw_model": node['hw_model'],
-                                "last_seen": node['last_seen']
-                            }
-                        }
-                    else:
-                        # No coords - create feature with null geometry
-                        feature = {
-                            "type": "Feature",
-                            "geometry": None,
-                            "properties": {
-                                "id": node['id'],
-                                "name": node['name'],
-                                "hw_model": node['hw_model'],
-                                "last_seen": node['last_seen'],
-                                "note": "No coordinates available"
-                            }
-                        }
-                    features.append(feature)
-
-                geojson = {
-                    "type": "FeatureCollection",
-                    "features": features,
-                    "properties": {
-                        "exported": datetime.now().isoformat(),
-                        "source": "MeshForge"
-                    }
-                }
-
-                # Save to file
-                export_dir = get_real_user_home() / ".local" / "share" / "meshforge" / "exports"
-                export_dir.mkdir(parents=True, exist_ok=True)
-                filename = f"nodes_{datetime.now().strftime('%Y%m%d_%H%M%S')}.geojson"
-                export_path = export_dir / filename
-
-                with open(export_path, 'w') as f:
-                    json.dump(geojson, f, indent=2)
-
-                nodes_with_coords = sum(1 for f in features if f['geometry'])
-                GLib.idle_add(self._log_message, f"Exported {len(nodes)} nodes ({nodes_with_coords} with coordinates)")
-                GLib.idle_add(self._log_message, f"Saved to: {export_path}")
-
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"Export error: {e}")
-
-        threading.Thread(target=do_export, daemon=True).start()
-
-    def _on_run_health_check(self, button):
-        """Run full health check"""
-        self._log_message("Running health check...")
-        self._update_health_cards()
-
-    def _on_ping_test(self, button):
-        """Run ping test"""
-        self._run_network_test("Ping Test", ['ping', '-c', '4', '8.8.8.8'])
-
-    def _on_traceroute_test(self, button):
-        """Run traceroute - try multiple tools"""
-        self._log_message("Running route trace...")
-
-        def do_trace():
-            # Try different traceroute tools in order of preference
-            trace_cmds = [
-                (['traceroute', '-m', '10', '8.8.8.8'], 'traceroute'),
-                (['mtr', '-r', '-c', '3', '8.8.8.8'], 'mtr'),
-                (['tracepath', '8.8.8.8'], 'tracepath'),
-            ]
-
-            for cmd, name in trace_cmds:
-                try:
-                    result = subprocess.run(cmd, capture_output=True, text=True, timeout=60)
-                    if result.returncode == 0 or result.stdout:
-                        output = result.stdout if result.stdout else result.stderr
-                        GLib.idle_add(self._log_message, f"\n=== {name} ===\n{output}")
-                        return
-                except FileNotFoundError:
-                    continue
-                except Exception as e:
-                    GLib.idle_add(self._log_message, f"{name} error: {e}")
-                    continue
-
-            GLib.idle_add(self._log_message, "No trace tools found. Install: sudo apt install traceroute")
-
-        threading.Thread(target=do_trace, daemon=True).start()
-
-    def _on_dns_test(self, button):
-        """Run DNS check - try multiple tools"""
-        self._log_message("Running DNS Check...")
-
-        def do_dns_test():
-            import socket
-            test_hosts = ['google.com', 'cloudflare.com', '1.1.1.1']
-            results = []
-
-            for host in test_hosts:
-                try:
-                    ip = socket.gethostbyname(host)
-                    results.append(f"  {host} -> {ip}")
-                except socket.gaierror as e:
-                    results.append(f"  {host} -> FAILED ({e})")
-
-            output = "=== DNS Resolution Test ===\n" + "\n".join(results)
-
-            # Also try to get nameservers from resolv.conf
-            try:
-                with open('/etc/resolv.conf', 'r') as f:
-                    resolv = f.read()
-                    nameservers = [line for line in resolv.split('\n') if line.startswith('nameserver')]
-                    if nameservers:
-                        output += "\n\n=== Nameservers ===\n" + "\n".join(nameservers)
-            except Exception:
-                pass
-
-            GLib.idle_add(self._log_message, output)
-
-        threading.Thread(target=do_dns_test, daemon=True).start()
-
-    def _on_port_scan(self, button):
-        """Run port scan on localhost"""
-        self._run_network_test("Port Scan", ['ss', '-tuln'])
-
-    def _run_network_test(self, name: str, cmd: list):
-        """Run a network test and show output"""
-        self._log_message(f"Running {name}...")
-
-        def do_test():
-            try:
-                result = subprocess.run(cmd, capture_output=True, text=True, timeout=30)
-                output = result.stdout if result.stdout else result.stderr
-                GLib.idle_add(self._log_message, f"\n=== {name} ===\n{output}")
-            except subprocess.TimeoutExpired:
-                GLib.idle_add(self._log_message, f"{name}: Timed out")
-            except Exception as e:
-                GLib.idle_add(self._log_message, f"{name} error: {e}")
-
-        threading.Thread(target=do_test, daemon=True).start()
-
-    def _on_generate_report(self, button):
-        """Generate diagnostic report"""
-        self._log_message("Generating diagnostic report...")
-
-        def do_report():
-            import socket
-            from datetime import datetime
-
-            report_lines = []
-            report_lines.append("=" * 60)
-            report_lines.append("MeshForge Diagnostic Report")
-            report_lines.append(f"Generated: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-            report_lines.append("=" * 60)
-
-            # System info
-            report_lines.append("\n[System Info]")
-            try:
-                uname = subprocess.run(['uname', '-a'], capture_output=True, text=True, timeout=5)
-                report_lines.append(f"  {uname.stdout.strip()}")
-            except Exception:  # Non-critical for report generation
-                pass
-
-            # MeshBot status
-            report_lines.append("\n[MeshBot Status]")
-            meshbot_path = self._path_entry.get_text().strip() or "/opt/meshing-around"
-            if Path(meshbot_path).exists():
-                report_lines.append(f"  Installed: Yes ({meshbot_path})")
-                venv = Path(meshbot_path) / "venv"
-                report_lines.append(f"  Venv: {'Yes' if venv.exists() else 'No'}")
-            else:
-                report_lines.append("  Installed: No")
-
-            # Check if running
-            try:
-                result = subprocess.run(['pgrep', '-f', 'mesh_bot.py'], capture_output=True, timeout=5)
-                report_lines.append(f"  Running: {'Yes' if result.returncode == 0 else 'No'}")
-            except Exception:
-                pass
-
-            # Network
-            report_lines.append("\n[Network]")
-            try:
-                ip = socket.gethostbyname('google.com')
-                report_lines.append(f"  DNS: Working ({ip})")
-            except Exception:
-                report_lines.append("  DNS: Failed")
-
-            # Meshtasticd
-            report_lines.append("\n[Meshtasticd]")
-            try:
-                result = subprocess.run(['systemctl', 'is-active', 'meshtasticd'], capture_output=True, text=True, timeout=5)
-                report_lines.append(f"  Service: {result.stdout.strip()}")
-            except Exception:
-                report_lines.append("  Service: Unknown")
-
-            # Check TCP port
-            sock = None
-            try:
-                sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-                sock.settimeout(2)
-                result = sock.connect_ex(('localhost', 4403))
-                report_lines.append(f"  TCP 4403: {'Open' if result == 0 else 'Closed'}")
-            except Exception:
-                pass
-            finally:
-                if sock:
-                    try:
-                        sock.close()
-                    except Exception:  # Ignore errors during cleanup
-                        pass
-
-            report_lines.append("\n" + "=" * 60)
-
-            # Display report
-            report_text = "\n".join(report_lines)
-            GLib.idle_add(self._set_log_text, report_text)
-            GLib.idle_add(self._log_message, "Report generated")
-
-        threading.Thread(target=do_report, daemon=True).start()
-
-    def _on_open_reports_folder(self, button):
-        """Open reports folder"""
-        reports_dir = get_real_user_home() / ".local" / "share" / "meshforge" / "diagnostics"
-        reports_dir.mkdir(parents=True, exist_ok=True)
-        self._open_folder(str(reports_dir))
+    # =========================================================================
+    # Log Management (Shared by all tabs)
+    # =========================================================================
 
     def _on_log_source_changed(self, combo):
         """Handle log source change"""
@@ -2143,6 +360,10 @@ class MeshToolsPanel(Gtk.Box):
         buffer = self._log_text.get_buffer()
         buffer.set_text(text)
 
+    # =========================================================================
+    # Utility Methods (Shared by all tabs)
+    # =========================================================================
+
     def _open_folder(self, path: str):
         """Open folder in file manager"""
         real_user = os.environ.get('SUDO_USER', os.environ.get('USER', 'pi'))
@@ -2194,6 +415,10 @@ class MeshToolsPanel(Gtk.Box):
             GLib.idle_add(self._log_message, "Copy URL and open manually in browser")
 
         threading.Thread(target=do_open, daemon=True).start()
+
+    # =========================================================================
+    # Cleanup
+    # =========================================================================
 
     def cleanup(self):
         """Clean up resources"""
